+++ AusSem/commune.h
@@ -6,14 +6,14 @@ class commune
 private:
 	char* name = nullptr;
 	char* code = nullptr;
-	size_t male = 0;
-	size_t female = 0;
+	unsigned int male = 0;
+	unsigned int female = 0;
 public:
 	commune(const  char* name, const  char* code, size_t male, size_t female);
 	commune(const commune& other);
 	commune& operator=(const commune& other);
 	~commune();
-	size_t getPopulation();
+	unsigned getPopulation();
 	void print();
 	
 	const char* getName();
